static_cast pixel.alpha

+++ cpp/spritefire/src/utils/color.cpp
@@ -9,5 +9,8 @@ RGBColor getRGB(Point point, png::image<png::rgba_pixel> image) {
 
 RGBAColor getRGBA(Point point, png::image<png::rgba_pixel> image) {
     png::rgba_pixel pixel = image[point.y][point.x];
-    return RGBAColor{RGBColor{pixel.red, pixel.green, pixel.blue}, pixel.alpha};
+    return RGBAColor{
+        RGBColor{pixel.red, pixel.green, pixel.blue}, 
+        static_cast<float>(pixel.alpha)
+    };
 }
